Editor client: Cleanup and split home component
