+++ crates/lgn-analytics-gui/frontend/package.json
@@ -28,29 +28,29 @@
     "@testing-library/jest-dom": "^5.16.2",
     "@testing-library/svelte": "^3.0.3",
     "@types/d3": "^7.1.0",
-    "@types/jest": "^27.4.0",
+    "@types/jest": "^27.4.1",
     "@typescript-eslint/eslint-plugin": "^5.12.1",
     "@typescript-eslint/parser": "^5.12.1",
     "autoprefixer": "^10.4.2",
     "eslint": "^8.9.0",
     "eslint-config-prettier": "^8.4.0",
-    "eslint-plugin-svelte3": "^3.4.0",
+    "eslint-plugin-svelte3": "^3.4.1",
     "jest": "^27.5.1",
-    "pnpm": "^6.31.0",
+    "pnpm": "^6.32.1",
     "postcss": "^8.4.6",
     "postcss-load-config": "^3.1.3",
     "prettier": "2.5.1",
     "prettier-plugin-svelte": "^2.6.0",
-    "puppeteer": "^13.3.2",
+    "puppeteer": "^13.4.0",
     "svelte": "^3.46.4",
     "svelte-check": "^2.4.5",
     "svelte-jester": "^2.3.2",
-    "svelte-preprocess": "^4.10.3",
+    "svelte-preprocess": "^4.10.4",
     "tailwindcss": "^3.0.23",
     "ts-proto": "^1.106.1",
     "typescript": "^4.5.5",
     "vite": "^2.8.4",
-    "vite-tsconfig-paths": "^3.4.0"
+    "vite-tsconfig-paths": "^3.4.1"
   },
   "dependencies": {
     "@improbable-eng/grpc-web": "^0.15.0",
