+++ crates/lgn-content-store-proto/package.json
@@ -15,7 +15,7 @@
   "devDependencies": {
     "jest": "^27.4.7",
     "ts-jest": "^27.1.3",
-    "ts-proto": "^1.104.0",
+    "ts-proto": "^1.106.1",
     "typescript": "^4.5.5"
   },
   "dependencies": {
@@ -23,7 +23,7 @@
     "browser-headers": "^0.4.1",
     "google-protobuf": "^3.19.4",
     "long": "^5.2.0",
-    "pnpm": "^6.29.1",
+    "pnpm": "^6.32.1",
     "protobufjs": "^6.11.2"
   }
 }
