+++ crates/lgn-editor-gui/frontend/src/components/resources/ResourceFilter.svelte
@@ -25,12 +25,21 @@
   class="flex items-center w-full space-x-1 justify-end py-1 px-2"
   on:submit={submit}
 >
-  <TextInput bind:value={name} size="default" fluid placeholder="Resource Name">
-    <div class="clear" slot="rightExtension" on:click={resetname}>
-      <Icon icon="ic:baseline-close" title="Reset filter" />
-    </div>
-  </TextInput>
-  <Button type="submit">Search</Button>
+  <div class="flex-grow">
+    <TextInput
+      bind:value={name}
+      size="default"
+      fluid
+      placeholder="Resource Name"
+    >
+      <div class="clear" slot="rightExtension" on:click={resetname}>
+        <Icon icon="ic:baseline-close" title="Reset filter" />
+      </div>
+    </TextInput>
+  </div>
+  <div class="flex-shrink">
+    <Button type="submit">Search</Button>
+  </div>
 </form>
 
 <style lang="postcss">
