+++ crates/lgn-editor-gui/frontend/src/index.ts
@@ -3,6 +3,9 @@ import "./assets/index.css";
 import { AppComponent, run } from "@lgn/web-client";
 import App from "@/App.svelte";
 import "@/workers/editorWorker";
+import contextMenu from "@/stores/contextMenu";
+import * as contextMenuEntries from "@/data/contextMenu";
+import viewportOrchestrator from "@/stores/viewport";
 //import initWasmLogger, { debug } from "@lgn/simple-wasm-logger";
 
 const redirectUri = document.location.origin + "/";
@@ -38,6 +41,21 @@ run({
   async onPreInit() {
     //await initWasmLogger();
     //debug("Hello from the Legion editor");
+
+    contextMenu.register("resource", contextMenuEntries.resourceEntries);
+    contextMenu.register(
+      "resourcePanel",
+      contextMenuEntries.resourcePanelEntries
+    );
+
+    const editorViewportKey = Symbol();
+
+    viewportOrchestrator.addAllViewport(
+      [editorViewportKey, { type: "video", name: "editor" }],
+      [Symbol(), { type: "video", name: "runtime" }]
+    );
+
+    viewportOrchestrator.activate(editorViewportKey);
   },
 })
   // eslint-disable-next-line no-console
