+++ crates/lgn-editor-gui/frontend/src/lib/hierarchyTree.ts
@@ -5,7 +5,6 @@ export type Entry<Item> = {
   index: number;
   item: Item;
   subEntries: Entry<Item>[] | null;
-  // eslint-disable-next-line @typescript-eslint/ban-types
 };
 
 // TODO: Improve performance if needed, and stop using recursion
@@ -13,7 +12,7 @@ export type Entry<Item> = {
 export class Entries<Item> {
   entries: Entry<Item>[];
 
-  #size: number | null = null;
+  #size!: number;
 
   /**
    * Build an `Entries` object from any flat arrays of object.
@@ -94,14 +93,12 @@ export class Entries<Item> {
   /** Builds an `Entries` object from and array of `Entry` */
   constructor(entries: Entry<Item>[]) {
     this.entries = entries;
+
+    this.recalculateSize();
   }
 
   /** Computes the size of the `Entries` */
-  get size(): number {
-    if (typeof this.#size === "number") {
-      return this.#size;
-    }
-
+  recalculateSize(): number {
     function count(entries: Entry<Item>[], size = 0): number {
       return entries.reduce(
         (size, entry) =>
@@ -171,8 +168,7 @@ export class Entries<Item> {
 
     this.entries = filter(this.entries);
 
-    // Resets the size so it's computed again if needed
-    this.#size = null;
+    this.recalculateSize();
 
     this.#setIndices();
 
@@ -269,7 +265,7 @@ export class Entries<Item> {
 
     this.#sort();
 
-    this.#size = null;
+    this.recalculateSize();
 
     return this;
   }
@@ -287,6 +283,24 @@ export class Entries<Item> {
 
   remove(removedEntry: Entry<Item>): this {
     return this.filter((entry) => entry !== removedEntry);
+  }
+
+  isEmpty() {
+    return this.#size === 0;
+  }
+
+  intoItems() {
+    const items = [];
+
+    for (const [, entry] of this) {
+      items.push(entry.item);
+    }
+
+    return items;
+  }
+
+  size() {
+    return this.#size;
   }
 
   [Symbol.iterator]() {
