+++ crates/lgn-editor-gui/frontend/src/stores/modal.ts
@@ -1,0 +1,3 @@
+import ModalStore from "@lgn/web-client/src/stores/modal";
+
+export default new ModalStore();
