+++ crates/lgn-web-client/frontend/package.json
@@ -25,13 +25,13 @@
     "@testing-library/jest-dom": "^5.16.2",
     "@testing-library/svelte": "^3.0.3",
     "@types/color-convert": "^2.0.0",
-    "@types/jest": "^27.4.0",
+    "@types/jest": "^27.4.1",
     "@typescript-eslint/eslint-plugin": "^5.12.1",
     "@typescript-eslint/parser": "^5.12.1",
     "autoprefixer": "^10.4.2",
     "eslint": "^8.9.0",
     "eslint-config-prettier": "^8.4.0",
-    "eslint-plugin-svelte3": "^3.4.0",
+    "eslint-plugin-svelte3": "^3.4.1",
     "jest": "^27.5.1",
     "postcss": "^8.4.6",
     "postcss-load-config": "^3.1.3",
@@ -40,7 +40,7 @@
     "svelte": "^3.46.4",
     "svelte-check": "^2.4.5",
     "svelte-jester": "^2.3.2",
-    "svelte-preprocess": "^4.10.3",
+    "svelte-preprocess": "^4.10.4",
     "tailwindcss": "^3.0.23",
     "typescript": "^4.5.5",
     "vite": "^2.8.4"
@@ -56,7 +56,7 @@
     "long": "^5.2.0",
     "monaco-editor": "0.32.1",
     "oauth-pkce": "^0.0.6",
-    "pnpm": "^6.31.0",
+    "pnpm": "^6.32.1",
     "protobufjs": "^6.11.2",
     "rxjs": "6.6.7"
   }
